added slider bar for monster progress

+++ Testovoe/Assets/Scripts/Wave.cs
@@ -4,11 +4,11 @@ public class Wave : MonoBehaviour
 {
     [SerializeField] private GameObject _enemyPrefab;
     [SerializeField] private float _spawnDelay;
-    [SerializeField] private float _amount;
+    [SerializeField] private int _amount;
 
     public float SpawnDelay => _spawnDelay;
 
     public GameObject EnemyPrefab => _enemyPrefab;
 
-    public float Amount => _amount;
+    public int Amount => _amount;
 }
